Update all acceptance tests to include config
