+++ ec/acc/testdata/deployment_basic_topology_config.tf
@@ -1,20 +1,23 @@
 resource "ec_deployment" "basic" {
-  name                   = "%s"
-  region                 = "%s"
-  version                = "%s"
-  
+  name    = "%s"
+  region  = "%s"
+  version = "%s"
+
   # TODO: Make this template ID dependent on the region.
   deployment_template_id = "aws-io-optimized-v2"
 
   elasticsearch {
     topology {
       instance_configuration_id = "aws.data.highio.i3"
-      memory_per_node = "1g"
+      memory_per_node           = "1g"
     }
   }
 
   kibana {
     topology {
+      config {
+        user_settings_yaml = "csp.warnLegacyBrowsers: true"
+      }
       instance_configuration_id = "aws.kibana.r5d"
     }
   }
